Move recursion to utility
